+++ CHANGELOG.md
@@ -16,7 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `put_in_container` and `set_in_container` helper methods.
+- Added the following helper methods:
+    - `in_container`
+    - `put_in_container`
+    - `set_in_container`
 
 ### Changed